+++ Utility/MailManager/base64.cpp
@@ -86,7 +86,7 @@ std::string base64_encode(unsigned char const* bytes_to_encode, size_t in_len)
 
 std::string base64_decode(std::string const& encoded_string) 
 {
-  int in_len = encoded_string.size();
+  int in_len = (int)(encoded_string.size());
   int i = 0, j = 0, in_ = 0;
   unsigned char char_array_4[4], char_array_3[3];
   std::string ret;
